YUK-9339 Convert 'Manipulate Results' option in STARS inventory to use the new inventory operation filtering and actions.
- Added remaining inventory functionality to new inventory page
- Removed link to old inventory page

+++ yukon-client/yukon-web/src/com/cannontech/web/stars/dr/operator/inventory/service/impl/DeleteInventoryHelper.java
@@ -3,12 +3,12 @@ package com.cannontech.web.stars.dr.operator.inventory.service.impl;
 import com.cannontech.common.bulk.collection.inventory.InventoryCollection;
 import com.cannontech.common.inventory.InventoryIdentifier;
 import com.cannontech.user.YukonUserContext;
+import com.cannontech.web.stars.dr.operator.inventory.service.CollectionBasedInventoryTask;
 import com.cannontech.web.stars.dr.operator.inventory.service.InventoryActionsHelper;
-import com.cannontech.web.stars.dr.operator.inventory.service.AbstractInventoryTask;
 
 public class DeleteInventoryHelper extends InventoryActionsHelper {
 
-    public class DeleteInventoryTask extends AbstractInventoryTask {
+    public class DeleteInventoryTask extends CollectionBasedInventoryTask {
         
         public DeleteInventoryTask(InventoryCollection collection, YukonUserContext context) {
             this.collection = collection;
@@ -20,10 +20,13 @@ public class DeleteInventoryHelper extends InventoryActionsHelper {
                 @Override
                 public void run() {
                     for (InventoryIdentifier inv : collection.getList()) {
+                        if (canceled) break;
                         try {
                             hardwareService.deleteHardware(context, true, inv.getInventoryId());
+                            successCount++;
                         } catch (Exception e) {
                             log.error("Unable to delete inventory: " + inv, e);
+                            failedCount++;
                         } finally {
                             completedItems ++;
                         }
